+++ ft_strlcat.c
@@ -6,17 +6,17 @@
 /*   By: icheca-g <icheca-g@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/21 14:37:40 by icheca-g          #+#    #+#             */
-/*   Updated: 2024/10/15 16:36:28 by icheca-g         ###   ########.fr       */
+/*   Updated: 2024/10/15 19:24:04 by icheca-g         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlcat(char *dest, const char *src, unsigned int size)
+size_t	ft_strlcat(char *dest, const char *src,size_t size)
 {
-	unsigned int	dest_len;
-	unsigned int	src_len;
-	unsigned int	i;
+	size_t	dest_len;
+	size_t	src_len;
+	size_t	i;
 
 	dest_len = ft_strlen(dest);
 	src_len = ft_strlen(src);
@@ -38,8 +38,8 @@ unsigned int	ft_strlcat(char *dest, const char *src, unsigned int size)
 // {
 // 	char dest[20] = "Hello, ";
 // 	char src[] = "World!";
-// 	unsigned int size = 20;
-// 	unsigned int length = ft_strlcat(dest, src, size);
+// 	size_t size = 20;
+// 	size_t length = ft_strlcat(dest, src, size);
 // 	printf("Destination: %s\n", dest);
 // 	printf("Length of destination: %u\n", length);
 // 	return 0;
